商品详情页

+++ shop-dao/src/main/java/com/masiis/shop/dao/platform/certificate/CertificateMapper.java
@@ -29,5 +29,5 @@ public interface CertificateMapper {
 
     String findById(Integer id);
 
-    List<CertificateInfo> getCertificatesByUser(Integer userId);
+    List<CertificateInfo> getCertificatesByUser(Long userId);
 }
